api/requirements.txt
cryptography==2.1.4
Flask==1.0
gunicorn==19.8.1
pycryptodomex==3.6.6
PyJWT==1.5.3
PyYAML==3.12